feat(controllers): reorganize config endpoints and add end_user management
- Migrate emotion config endpoints to centralized memory_config_controller
- Create new end_user_controller for end-user management operations
- Create new memory_config_controller consolidating all memory configuration endpoints
- Refactor emotion_config_controller to retain only shared data models
- Update controller imports and router registration in __init__.py
- Standardize config endpoint naming (read_config, update_config patterns)
- Consolidate configuration management into unified memory_config module for improved maintainability
