+++ api/app/controllers/__init__.py
@@ -15,6 +15,7 @@ from . import (
     document_controller,
     emotion_config_controller,
     emotion_controller,
+    end_user_controller,
     file_controller,
     file_storage_controller,
     home_page_controller,
@@ -28,6 +29,7 @@ from . import (
     mcp_market_config_controller,
     memory_agent_controller,
     memory_analytics_controller,
+    memory_config_controller,
     memory_controller,
     memory_dashboard_controller,
     memory_episodic_controller,
@@ -85,6 +87,8 @@ manager_router.include_router(upload_controller.router)
 manager_router.include_router(memory_agent_controller.router)
 manager_router.include_router(memory_controller.router)
 manager_router.include_router(memory_analytics_controller.router)
+manager_router.include_router(memory_config_controller.router)
+manager_router.include_router(end_user_controller.router)
 manager_router.include_router(memory_dashboard_controller.router)
 manager_router.include_router(memory_storage_controller.router)
 manager_router.include_router(user_memory_controllers.router)
