+++ api/app/controllers/memory_analytics_controller.py
@@ -24,7 +24,6 @@ from app.core.response_utils import fail, success
 from app.db import get_db
 from app.dependencies import get_current_user
 from app.models.user_model import User
-from app.repositories.end_user_repository import EndUserRepository
 from app.repositories.workspace_repository import WorkspaceRepository
 from app.schemas.memory_storage_schema import GenerateCacheRequest
 from app.schemas.response_schema import ApiResponse
@@ -337,53 +336,7 @@ async def get_community_graph_data_api(
         return fail(BizCode.INTERNAL_ERROR, "社区图谱查询失败", str(e))
 
 
-@router.get("/end_user_info", response_model=ApiResponse)
-async def get_end_user_info(
-    end_user_id: str,
-    current_user: User = Depends(get_current_user),
-    db: Session = Depends(get_db),
-) -> dict:
-    """
-    查询终端用户信息记录
-
-    根据 end_user_id 查询单条终端用户信息记录。
-    """
-    workspace_id = current_user.current_workspace_id
-
-    if workspace_id is None:
-        api_logger.warning(f"用户 {current_user.username} 尝试查询终端用户信息但未选择工作空间")
-        return fail(BizCode.INVALID_PARAMETER, "请先切换到一个工作空间", "current_workspace_id is None")
-
-    api_logger.info(
-        f"查询终端用户信息请求: end_user_id={end_user_id}, user={current_user.username}, "
-        f"workspace={workspace_id}"
-    )
-
-    end_user_repo = EndUserRepository(db)
-    end_user = end_user_repo.get_end_user_by_id(end_user_id)
-    if end_user is None:
-        return fail(BizCode.USER_NOT_FOUND, "终端用户不存在", "end_user not found")
-    if str(end_user.workspace_id) != str(workspace_id):
-        api_logger.warning(
-            f"用户 {current_user.username} 尝试查询不属于工作空间 {workspace_id} 的终端用户 {end_user_id}"
-        )
-        return fail(BizCode.PERMISSION_DENIED, "该终端用户不属于当前工作空间", "end_user workspace mismatch")
-
-    result = user_memory_service.get_end_user_info(db, end_user_id)
-
-    if result["success"]:
-        api_logger.info(f"成功查询终端用户信息: end_user_id={end_user_id}")
-        return success(data=result["data"], msg="查询成功")
-    else:
-        error_msg = result["error"]
-        api_logger.error(f"查询终端用户信息失败: end_user_id={end_user_id}, error={error_msg}")
-
-        if error_msg == "终端用户信息记录不存在":
-            return fail(BizCode.USER_NOT_FOUND, "终端用户信息记录不存在", error_msg)
-        elif error_msg == "无效的终端用户ID格式":
-            return fail(BizCode.INVALID_USER_ID, "无效的终端用户ID格式", error_msg)
-        else:
-            return fail(BizCode.INTERNAL_ERROR, "查询终端用户信息失败", error_msg)
+# get_end_user_info 已迁移至 end_user_controller（/end_user/info）
 
 
 @router.get("/interest_distribution", response_model=ApiResponse)
